chore: update docstrings
